criando o enum Vinculo
