+++ generateData/enums/vinculo.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Vinculo(Enum):
+    ESTUDANTE = 1,
+    COLABORADOR = 2
